requirements.txt
streamlit
google-generativeai
chromadb
PyMuPDF
tiktoken
python-dotenv
matplotlib
plotly